fix FS router circular dep by moving import
