+++ .changeset/funny-spies-care.md
@@ -1,0 +1,5 @@
+---
+"@solidjs/start": minor
+---
+
+fix FS router circular dep by moving import
