+++ examples/hackernews/src/app.tsx
@@ -1,5 +1,5 @@
 import { Router } from "@solidjs/router";
-import { FileRoutes } from "@solidjs/start";
+import { FileRoutes } from "@solidjs/start/router";
 import { Suspense } from "solid-js";
 import "./app.css";
 import Nav from "./components/nav";
