+++ examples/with-solid-styled/src/app.tsx
@@ -1,6 +1,6 @@
 import { MetaProvider } from "@solidjs/meta";
 import { Router } from "@solidjs/router";
-import { FileRoutes } from "@solidjs/start";
+import { FileRoutes } from "@solidjs/start/router";
 import { Suspense } from "solid-js";
 
 import { useAssets } from "solid-js/web";
