+++ examples/with-unocss/src/app.tsx
@@ -2,7 +2,7 @@ import "@unocss/reset/tailwind.css";
 import "virtual:uno.css";
 
 import { Router } from "@solidjs/router";
-import { FileRoutes } from "@solidjs/start";
+import { FileRoutes } from "@solidjs/start/router";
 import { Suspense } from "solid-js";
 import Nav from "~/components/Nav";
