+++ packages/start/client/index.ts
@@ -1,5 +1,3 @@
-// @refresh skip
-// TODO rename to index.tsx?
 import "vinxi/client";
 export { StartClient } from "./StartClient";
 export { mount } from "./mount";
