+++ packages/start/client/spa/index.ts
