+++ packages/start/router/index.ts
@@ -1,0 +1,1 @@
+export { FileRoutes } from "./FileRoutes";
