+++ packages/start/server/spa/index.ts
@@ -1,6 +1,4 @@
-// @refresh skip
-// TODO rename to index.ts?
+export * from "../types";
 export { StartServer } from "./StartServer";
 export { createHandler } from "./handler";
-export * from "./types";
