+++ packages/start/shared/index.ts
@@ -1,6 +1,3 @@
-// @refresh skip
-// TODO rename to index.ts?
-export { FileRoutes } from "./FileRoutes";
 export { GET } from "./GET";
 export { HttpHeader } from "./HttpHeader";
 export { HttpStatusCode } from "./HttpStatusCode";
